Refactor widget imports; move switch_button and app_icon imports to the app feature directory for improved organization and maintainability.
