+++ lib/ui/features/assets/widgets/tree/node/asset_status_icon.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 import '../../../../../../core/domain/company_asset/statuses.dart';
-import '../../../../../widgets/app_icon.dart';
+import '../../../../app/widgets/app_icon.dart';
 
 class AssetStatusIcon extends StatelessWidget {
   const AssetStatusIcon({super.key, required this.status});
